requirements.txt
absl-py==1.4.0
fse==1.0.0
lime==0.2.0.1
matplotlib==3.8.0
nltk==3.8.1
numpy==1.26.0
pandas==2.1.1
rbo==0.1.3
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
stanza==1.6.1
tensorflow==2.10.0
tensorflow-hub==0.15.0
torch==2.0.1
tqdm==4.66.1
ipykernel
unzip

